+++ CHANGELOG.md
@@ -13,7 +13,8 @@ feat: add `ADMIN_TABLE_DELETE_ROW` and `ADMIN_TABLE_INSERT_ROW` permissions
 
 ### `Changed` for changes in existing functionality
 
-feat: refer to jquery/3.7.1 in BlueprintWeb.latte always
+- feat: refer to jquery/3.7.1 in BlueprintWeb.latte always
+- feat: use various permissions to AdminHelper::getAllowedTables
 
 ### `Deprecated` for soon-to-be removed features